fix: make the JIT and loop passes agree with the interpreter on fuzzed programs
Both fixed-range fuzz ctests now pass and wide fuzzing (tens of thousands of
seeds across the bronze and all-passes pipelines) is clean.

x64 JIT: switch-case args uncounted in use analysis (flaky stack reads);
udiv/umod by constants never computed; srem by 2^k>=32 mask truncated;
address folding dropping still-live values; loads fused into ops that don't
use them; 32-bit spills leaving stale upper halves; edge/trampoline/deopt
blocks missing from the CFG; peephole store->load forwarding and copy-prop
unsound; NaN-unaware fused float compares and f32 via ucomisd; parallel-copy
cycle clobber and imm64 truncation; GC roots at safepoints computed from
block order instead of liveness; self mov32 (zero-extend) dropped; lea
absorbing an add that fed a folded address. Missing registers are now hard
errors in rewrite, emit and isel.

GVN-PRE: commutative reordering of ptr+i64; a same-address store treated as
transparent, hoisting loads out of loops that store to them.

Loops: parallel-loop analysis and transform rewritten to refuse unprovable
shapes (branches in bodies, exit values, reduction init, exit polarity,
runtime steps behind a wrap guard, vector accesses, unaliased bases, offset
gaps, overflow UB in reductions); tiling rewritten as in-place outer-two
tiling (unsound matmul interchange removed); unroll accumulator splitting
bails when another value reads the running sum; passes iterate a snapshot
of the function list.

AArch64 and PTX: integer division traps on zero and wraps MIN/-1 as MIR
defines. Parser split into parser_impl/parser_function/parser.

+++ include/brass/runtime/parallel_runtime.hpp
@@ -45,8 +45,11 @@ inline bool is_fp_reduction(ReductionKind k) noexcept {
 
 inline int64_t combine_reduction_i64(ReductionKind kind, int64_t acc, int64_t val) noexcept {
     switch (kind) {
-        case ReductionKind::SumI64: return acc + val;
-        case ReductionKind::ProdI64: return acc * val;
+        // MIR integer arithmetic wraps; combine in unsigned to match it.
+        case ReductionKind::SumI64:
+            return static_cast<int64_t>(static_cast<uint64_t>(acc) + static_cast<uint64_t>(val));
+        case ReductionKind::ProdI64:
+            return static_cast<int64_t>(static_cast<uint64_t>(acc) * static_cast<uint64_t>(val));
         case ReductionKind::MinI64: return std::min(acc, val);
         case ReductionKind::MaxI64: return std::max(acc, val);
         default: return acc;
